helper: split Nar(Info)Path into Nar(Info)FilePath and Nar(Info)URLPath (#64)
I will need to implement sharding, and the best way to do it is within
the helper. However, sharding should only happen for store paths and not
for urls. By splitting the function now, I can change the functions and
fix tests without having to worry about if the call was correct in the
first place.

ref #63