pkg/cache: Save compressed NAR when compression is only over HTTP (#108)
If the upstream server returns a `narinfo` that has `Compression: none` then set `Accept-Encoding: zstd` and make an immediate request for the nar. If the server returns a response with the header `Content-Encoding: zstd` set, then save the response as-is in the store (without decompressing it) and modify the narinfo (update `Compression`, `URL` and `FileSize`) before storing it.

This will add specific support for [Harmonia](https://github.com/nix-community/harmonia) since it always returns nars uncompressed, but will compress them during transport at the HTTP layer.

closes #62